requirements.txt
python-dotenv==1.0.0
beautifulsoup4==4.12.2
agno